+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-09-23] [Add onKeyDown prop to Input component and update InputProps type definition](https://github.com/RubricLab/fabric-ui/commit/9aa891b91b317c52cac64a454a9d4f7df731a084)
 - [2025-09-22] [Add value prop to Select component and update SelectProps type definition to include value](https://github.com/RubricLab/fabric-ui/commit/29119aec60e8a6bd211fa9820f99a2c9dff4d456)
 - [2025-09-18] [Add onKeyDown prop to Form component and update FormProps type definition](https://github.com/RubricLab/fabric-ui/commit/36ae14ec97fc1f0279b37d1b82fdcfda56ca480a)
 - [2025-09-17] [Refactor components to use rounded utility classes for improved styling consistency](https://github.com/RubricLab/fabric-ui/commit/05c166223733cfe93b5052a0b119d94f1ae5bc69)